This works until we create additional Firms. Problem seems to be that closing balance (Producers sector) doesn't keep track of purchases that are made by workers that are not employees of the firm from which they are purchasing. Solution is probably to remove prod_bal from stats and iterate through a new Firm::getBalance() method at appropriate point in the cycle.

+++ Test Product/statistics.hpp
@@ -13,6 +13,8 @@
 #include <iostream>
 #include <vector>
 
+#include "account.hpp"
+
 class Statistics
 {
 private:
@@ -21,8 +23,8 @@ private:
 
     struct fields
     {
-        int num_firms = 0;
-        int num_employed = 0;
+        size_t num_firms = 0;
+        size_t num_employed = 0;
         int num_hired = 0;
         int num_fired = 0;
         int num_rehired = 0;
